+++ src/main/java/dev/masa/masuiteteleports/core/objects/TeleportType.java
@@ -1,0 +1,8 @@
+package dev.masa.masuiteteleports.core.objects;
+
+public enum TeleportType {
+    FORCE_HERE,
+    FORCE_TO,
+    REQUEST_HERE,
+    REQUEST_TO
+}
