--- src/main/java/fi/matiaspaavilainen/masuiteteleports/core/objects/TeleportType.java
@@ -1,8 +1,0 @@
-package fi.matiaspaavilainen.masuiteteleports.core.objects;
-
-public enum TeleportType {
-    FORCE_HERE,
-    FORCE_TO,
-    REQUEST_HERE,
-    REQUEST_TO
-}
